Fix RST parsing in schemagen
- Use Python standardized docstring "trim" function to preserve newlines
  and indentation in schelp strings.
- Fix schemagen parse_rst() to handle multiline text
