+++ docs/_ext/dynamicgen.py
@@ -19,7 +19,7 @@ import subprocess
 
 from common import *
 
-import siliconcompiler
+from siliconcompiler import utils
 
 #############
 # Helpers
@@ -98,37 +98,6 @@ def build_config_recursive(schema, keypath_prefix=[], sec_key_prefix=[]):
         # schema without leaves.
         return child_sections
 
-def trim(docstring):
-    '''Helper function for cleaning up indentation of docstring.
-
-    This is important for properly parsing complex RST in our docs.
-
-    Source:
-    https://www.python.org/dev/peps/pep-0257/#handling-docstring-indentation'''
-    if not docstring:
-        return ''
-    # Convert tabs to spaces (following the normal Python rules)
-    # and split into a list of lines:
-    lines = docstring.expandtabs().splitlines()
-    # Determine minimum indentation (first line doesn't count):
-    indent = sys.maxsize
-    for line in lines[1:]:
-        stripped = line.lstrip()
-        if stripped:
-            indent = min(indent, len(line) - len(stripped))
-    # Remove indentation (first line is special):
-    trimmed = [lines[0].strip()]
-    if indent < sys.maxsize:
-        for line in lines[1:]:
-            trimmed.append(line[indent:].rstrip())
-    # Strip off trailing and leading blank lines:
-    while trimmed and not trimmed[-1]:
-        trimmed.pop()
-    while trimmed and not trimmed[0]:
-        trimmed.pop(0)
-    # Return a single string:
-    return '\n'.join(trimmed)
-
 #############
 # Base class
 #############
@@ -159,7 +128,7 @@ class DynamicGen(SphinxDirective):
         # raw docstrings have funky indentation (basically, each line is already
         # indented as much as the function), so we call trim() helper function
         # to clean it up
-        docstr = trim(make_docs.__doc__)
+        docstr = utils.trim(make_docs.__doc__)
 
         if docstr:
             self.parse_rst(docstr, s)
